+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Release History
 
+## 0.3.0
+* New API
+** remove `compareTo()` returning 0, -1 and -1
+** add `equals()` returning true or false
+
 ## 0.2.0
 * New API
 ** only accept array as constructor parameter